requirements.txt
requests==2.28.2
selenium-4.8.2==4.8.2
pre-commit==3.2.0
pandas==1.5.3
